requirements.txt
amqp==2.5.1
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
billiard==3.5.0.5
celery==4.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
Django==2.0.5
django-elasticsearch-dsl==7.0.0
django-haystack==2.8.1
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
Faker==0.9.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.22
kombu==4.1.0
mixer==6.1.3
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.1
pluggy==0.13.0
postgres==2.2.2
psycopg2==2.7.4
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
text-unidecode==1.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
wikipedia==1.4.0
zipp==0.6.0
